refactor: 유틸 함수 정리, 부모 클래스 생성
