+++ src/shared/component.js
@@ -1,0 +1,11 @@
+export default class Component {
+  constructor($root, params) {
+    this.$root = $root;
+    this.params = params;
+  }
+
+  init() {}
+  getTemplate() {}
+  render() {}
+  addEvent() {}
+}
